+++ app.vue
@@ -3,3 +3,9 @@
 		<NuxtPage />
 	</NuxtLayout>
 </template>
+
+<script setup lang="ts">
+useHead({
+	title: 'Hair Recovery',
+})
+</script>
