+++ components/TheNavbar.vue
@@ -7,8 +7,6 @@
 	</nav>
 </template>
 <script setup lang="ts">
-import { useRoute, useRouter } from 'vue-router'
-
 const route = useRoute()
 const router = useRouter()
