+++ components/base/BaseCard.vue
@@ -1,0 +1,5 @@
+<template>
+	<div class="bg-white shadow-lg rounded-lg overflow-hidden p-6">
+		<slot />
+	</div>
+</template>
