+++ helpers/config.ts
@@ -1,0 +1,1 @@
+export const PrefixTitle = 'HR -'
